move fd casting into readEventTimeout

+++ uapi/uapi_test.go
@@ -1618,30 +1618,30 @@ func TestReadEvent(t *testing.T) {
 	require.Nil(t, err)
 	defer unix.Close(int(er.Fd))
 
-	evt, err := readEventTimeout(uintptr(er.Fd), spuriousEventWaitTimeout)
+	evt, err := readEventTimeout(er.Fd, spuriousEventWaitTimeout)
 	assert.Nil(t, err)
 	assert.Nil(t, evt, "spurious event")
 
 	c.SetValue(1, 1)
-	evt, err = readEventTimeout(uintptr(er.Fd), eventWaitTimeout)
+	evt, err = readEventTimeout(er.Fd, eventWaitTimeout)
 	require.Nil(t, err)
 	require.NotNil(t, evt)
 	assert.Equal(t, uapi.EventRequestFallingEdge, evt.ID)
 
 	c.SetValue(1, 0)
-	evt, err = readEventTimeout(uintptr(er.Fd), eventWaitTimeout)
+	evt, err = readEventTimeout(er.Fd, eventWaitTimeout)
 	assert.Nil(t, err)
 	require.NotNil(t, evt)
 	assert.Equal(t, uapi.EventRequestRisingEdge, evt.ID)
 }
 
-func readEventTimeout(fd uintptr, t time.Duration) (*uapi.EventData, error) {
+func readEventTimeout(fd int32, t time.Duration) (*uapi.EventData, error) {
 	pollfd := unix.PollFd{Fd: int32(fd), Events: unix.POLLIN}
 	n, err := unix.Poll([]unix.PollFd{pollfd}, int(t.Milliseconds()))
 	if err != nil || n != 1 {
 		return nil, err
 	}
-	evt, err := uapi.ReadEvent(fd)
+	evt, err := uapi.ReadEvent(uintptr(fd))
 	if err != nil {
 		return nil, err
 	}
